+++ ios/Podfile
@@ -79,6 +79,8 @@ target 'SecretariaVirtualTotem' do
 
   pod 'react-native-orientation-locker', :path => '../node_modules/react-native-orientation-locker'
 
+  pod 'RNGestureHandler', :path => '../node_modules/react-native-gesture-handler'
+
   target 'SecretariaVirtualTotemTests' do
     inherit! :complete
     # Pods for testing
