+++ navigation/AppNavigator.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { createAppContainer, createSwitchNavigator } from 'react-navigation';
+
+import DrawerNavigator from './DrawerNavigator';
+
+export default createAppContainer(
+  createSwitchNavigator({
+    Main: DrawerNavigator
+  })
+);
