+++ package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "@react-native-async-storage/async-storage": "^1.13.1",
     "@react-native-community/masked-view": "^0.1.10",
-    "@react-navigation/drawer": "^5.9.3",
+    "@react-navigation/drawer": "^5.10.2",
+    "@react-navigation/native": "^5.8.2",
     "axios": "^0.20.0",
     "formik": "^2.1.5",
     "react": "16.11.0",
@@ -25,12 +26,13 @@
     "react-native-reanimated": "^1.13.1",
     "react-native-responsive-screen": "^1.4.1",
     "react-native-safe-area-context": "^3.1.8",
-    "react-native-screens": "^2.11.0",
+    "react-native-screens": "^2.12.0",
     "react-native-simple-card-view": "^0.3.0",
     "react-native-size-matters": "^0.3.1",
     "react-native-svg": "^12.1.0",
     "react-native-vector-icons": "^7.1.0",
-    "react-navigation": "^4.4.0",
+    "react-navigation": "^4.4.3",
+    "react-navigation-drawer": "^2.6.0",
     "react-navigation-stack": "^2.8.2",
     "yup": "^0.29.3"
   },
